+++ lib/routes/preferences/sections/privacy.dart
@@ -1,6 +1,6 @@
 import "package:flow/l10n/extensions.dart";
-import "package:flow/prefs/local_preferences.dart";
 import "package:flow/routes/preferences_page.dart";
+import "package:flow/services/user_preferences.dart";
 import "package:flutter/material.dart";
 import "package:material_symbols_icons/symbols.dart";
 
@@ -14,20 +14,40 @@ class Privacy extends StatefulWidget {
 class _PrivacyState extends State<Privacy> {
   @override
   Widget build(BuildContext context) {
-    final bool privacyMode = LocalPreferences().privacyMode.get();
-
-    return SwitchListTile(
-      secondary: const Icon(Symbols.password_rounded),
-      title: Text("preferences.privacy.maskAtStartup".t(context)),
-      value: privacyMode,
-      onChanged: updatePrivacyMode,
+    return Column(
+      mainAxisSize: .min,
+      children: [
+        SwitchListTile(
+          secondary: const Icon(Symbols.password_rounded),
+          title: Text("preferences.privacy.maskAtStartup".t(context)),
+          value: UserPreferencesService().privacyModeUponLaunch,
+          onChanged: updatePrivacyMode,
+        ),
+        SwitchListTile(
+          secondary: const Icon(Symbols.earthquake_rounded),
+          title: Text("preferences.privacy.maskAtShake".t(context)),
+          value: UserPreferencesService().privacyModeUponShaking,
+          onChanged: updatePrivacyModeUponShaking,
+        ),
+      ],
     );
   }
 
   void updatePrivacyMode(bool? newPrivacyMode) async {
     if (newPrivacyMode == null) return;
 
-    await LocalPreferences().privacyMode.set(newPrivacyMode);
+    UserPreferencesService().privacyModeUponLaunch = newPrivacyMode;
+
+    if (!mounted) return;
+
+    PreferencesPage.of(context).reload();
+    setState(() {});
+  }
+
+  void updatePrivacyModeUponShaking(bool? newPrivacyMode) async {
+    if (newPrivacyMode == null) return;
+
+    UserPreferencesService().privacyModeUponShaking = newPrivacyMode;
 
     if (!mounted) return;